+++ src/Events/Messenger/Serializer.php
@@ -1,0 +1,229 @@
+<?php declare(strict_types=1);
+
+namespace Somnambulist\Domain\Events\Messenger;
+
+use Somnambulist\Domain\Events\AbstractDomainEvent;
+use Symfony\Component\Messenger\Envelope;
+use Symfony\Component\Messenger\Exception\InvalidArgumentException;
+use Symfony\Component\Messenger\Exception\LogicException;
+use Symfony\Component\Messenger\Exception\MessageDecodingFailedException;
+use Symfony\Component\Messenger\Stamp\NonSendableStampInterface;
+use Symfony\Component\Messenger\Stamp\SerializerStamp;
+use Symfony\Component\Messenger\Transport\Serialization\PhpSerializer;
+use Symfony\Component\Messenger\Transport\Serialization\SerializerInterface;
+use Symfony\Component\Serializer\Encoder\JsonEncoder;
+use Symfony\Component\Serializer\Encoder\XmlEncoder;
+use Symfony\Component\Serializer\Exception\UnexpectedValueException;
+use Symfony\Component\Serializer\Normalizer\ArrayDenormalizer;
+use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
+use Symfony\Component\Serializer\Serializer as SymfonySerializer;
+use Symfony\Component\Serializer\SerializerInterface as SymfonySerializerInterface;
+use function array_merge;
+use function class_exists;
+use function get_class;
+use function sprintf;
+use function strlen;
+use function strpos;
+use function substr;
+
+/**
+ * Class Serializer
+ *
+ * Custom Serializer derived from `Symfony\Component\Messenger\Transport\Serialization\Serializer` that
+ * can encode/decode the custom AbstractDomainEvent. The differences here are the use of the toArray /
+ * fromArray on the AbstractDomainEvent to avoid needing to use the SF Serializer.
+ *
+ * @package Somnambulist\Domain\Events\Messenger
+ * @subpackage Somnambulist\Domain\Events\Messenger\Serializer
+ */
+class Serializer implements SerializerInterface
+{
+
+    private const STAMP_HEADER_PREFIX = 'X-Message-Stamp-';
+
+    /**
+     * @var SymfonySerializer
+     */
+    private $serializer;
+
+    /**
+     * @var string
+     */
+    private $format;
+
+    /**
+     * @var array
+     */
+    private $context;
+
+    /**
+     * Constructor.
+     *
+     * @param SymfonySerializerInterface|null $serializer
+     * @param string                          $format
+     * @param array                           $context
+     */
+    public function __construct(SymfonySerializerInterface $serializer = null, string $format = 'json', array $context = [])
+    {
+        $this->serializer = $serializer ?? self::create()->serializer;
+        $this->format     = $format;
+        $this->context    = $context;
+    }
+
+    public static function create(): self
+    {
+        if (!class_exists(SymfonySerializer::class)) {
+            throw new LogicException(
+                sprintf(
+                    'The "%s" class requires Symfony\'s Serializer component. Try running "composer require symfony/serializer" or use "%s" instead.',
+                    __CLASS__,
+                    PhpSerializer::class
+                )
+            );
+        }
+
+        $encoders    = [new XmlEncoder(), new JsonEncoder()];
+        $normalizers = [new ArrayDenormalizer(), new ObjectNormalizer()];
+        $serializer  = new SymfonySerializer($normalizers, $encoders);
+
+        return new self($serializer);
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function decode(array $encodedEnvelope): Envelope
+    {
+        if (empty($encodedEnvelope['body']) || empty($encodedEnvelope['headers'])) {
+            throw new MessageDecodingFailedException('Encoded envelope should have at least a "body" and some "headers".');
+        }
+
+        if (empty($encodedEnvelope['headers']['type'])) {
+            throw new MessageDecodingFailedException('Encoded envelope does not have a "type" header.');
+        }
+
+        $stamps          = $this->decodeStamps($encodedEnvelope);
+        $serializerStamp = $this->findFirstSerializerStamp($stamps);
+
+        $context = $this->context;
+        if (null !== $serializerStamp) {
+            $context = $serializerStamp->getContext() + $context;
+        }
+
+        try {
+            $message = AbstractDomainEvent::fromArray(
+                $encodedEnvelope['headers']['type'],
+                $this->serializer->decode($encodedEnvelope['body'], $this->format, $context)
+            );
+
+        } catch (UnexpectedValueException $e) {
+            throw new MessageDecodingFailedException(sprintf('Could not decode message: %s.', $e->getMessage()), $e->getCode(), $e);
+        }
+
+        return new Envelope($message, $stamps);
+    }
+
+    private function decodeStamps(array $encodedEnvelope): array
+    {
+        $stamps = [];
+
+        foreach ($encodedEnvelope['headers'] as $name => $value) {
+            if (0 !== strpos($name, self::STAMP_HEADER_PREFIX)) {
+                continue;
+            }
+
+            try {
+                $stamps[] = $this->serializer->deserialize($value, substr($name, strlen(self::STAMP_HEADER_PREFIX)) . '[]', $this->format, $this->context);
+            } catch (UnexpectedValueException $e) {
+                throw new MessageDecodingFailedException(sprintf('Could not decode stamp: %s.', $e->getMessage()), $e->getCode(), $e);
+            }
+        }
+
+        if ($stamps) {
+            $stamps = array_merge(...$stamps);
+        }
+
+        return $stamps;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
+    public function encode(Envelope $envelope): array
+    {
+        if (!$envelope->getMessage() instanceof AbstractDomainEvent) {
+            throw new InvalidArgumentException(
+                sprintf('"%s" will only operate on "%s" objects', __CLASS__, AbstractDomainEvent::class)
+            );
+        }
+
+        $context = $this->context;
+        /** @var SerializerStamp|null $serializerStamp */
+        if ($serializerStamp = $envelope->last(SerializerStamp::class)) {
+            $context = $serializerStamp->getContext() + $context;
+        }
+
+        $envelope = $envelope->withoutStampsOfType(NonSendableStampInterface::class);
+
+        $headers = ['type' => get_class($envelope->getMessage())] + $this->encodeStamps($envelope) + $this->getContentTypeHeader();
+
+        return [
+            'body'    => $this->serializer->encode($envelope->getMessage()->toArray(), $this->format, $context),
+            'headers' => $headers,
+        ];
+    }
+
+    private function encodeStamps(Envelope $envelope): array
+    {
+        if (!$allStamps = $envelope->all()) {
+            return [];
+        }
+
+        $headers = [];
+        foreach ($allStamps as $class => $stamps) {
+            $headers[self::STAMP_HEADER_PREFIX . $class] = $this->serializer->serialize($stamps, $this->format, $this->context);
+        }
+
+        return $headers;
+    }
+
+    /**
+     * @param array|SerializerStamp[] $stamps
+     *
+     * @return SerializerStamp|null
+     */
+    private function findFirstSerializerStamp(array $stamps): ?SerializerStamp
+    {
+        foreach ($stamps as $stamp) {
+            if ($stamp instanceof SerializerStamp) {
+                return $stamp;
+            }
+        }
+
+        return null;
+    }
+
+    private function getContentTypeHeader(): array
+    {
+        $mimeType = $this->getMimeTypeForFormat();
+
+        return null === $mimeType ? [] : ['Content-Type' => $mimeType];
+    }
+
+    private function getMimeTypeForFormat(): ?string
+    {
+        switch ($this->format) {
+            case 'json':
+                return 'application/json';
+            case 'xml':
+                return 'application/xml';
+            case 'yml':
+            case 'yaml':
+                return 'application/x-yaml';
+            case 'csv':
+                return 'text/csv';
+        }
+
+        return null;
+    }
+}